// src/types/dashboard.ts
export interface DataRKAP {
  periode: string;
  id_project: string;
  project_name: string;
  pu_rkap: number;
  pu_real: number;
  persentase: number;
  segmentasi: string;
  klien: string;
  nilai_kontrak: number;
  progress_fisik: number;
  status_proyek: string;
  kota: string;
  lng: number;
  lat: number;
  has_koordinat: boolean;
}